Refactor authentication logic by creating reusable signIn function and adding user type handling in signUp process; update .gitignore to exclude package-lock.json
