+++ .gitignore
@@ -1,0 +1,1 @@
+frontend/package-lock.json
